Minor format edits and error removal from -Werror flag

+++ AlgorithmProblem.cpp
@@ -23,16 +23,14 @@ void AlgorithmProblem::Init(){
 	read_residents_file(residence_data_filepath);
 }
 
-void AlgorithmProblem::read_hospital_file(std::string hospital_filepath)
-{
+void AlgorithmProblem::read_hospital_file(std::string hospital_filepath) {
 	int index{ 1 };
 
 	std::ifstream hos_file;
 	hos_file.open(hospital_filepath);
 
 	std::string line;
-	while (std::getline(hos_file, line))
-	{
+	while (std::getline(hos_file, line)) {
 		std::istringstream iss(line);
 		Hospitals.emplace(index,
 			std::vector<int>(std::istream_iterator<int>(iss),
@@ -45,23 +43,18 @@ void AlgorithmProblem::read_hospital_file(std::string hospital_filepath)
 	hos_file.close();
 }
 
-void AlgorithmProblem::read_residents_file(std::string residences_filepath)
-{
+void AlgorithmProblem::read_residents_file(std::string residences_filepath) {
 	int index{ 1 };
 
 	std::ifstream res_file;
 	res_file.open(residences_filepath);
 
 	std::string line;
-	while (std::getline(res_file, line))
-	{
+	while (std::getline(res_file, line)) {
 		std::istringstream iss(line);
 		Residents.emplace(index, 
 			std::vector<int>(std::istream_iterator<int>(iss),
 				std::istream_iterator<int>()));
-		// ResidentsPreference.emplace(index, 
-		// 	std::vector<int>(std::istream_iterator<int>(iss),
-		// 		std::istream_iterator<int>()));
 		FreeResidents.push_back(index);
 		++index;
 	}
@@ -73,26 +66,12 @@ void AlgorithmProblem::printValues(){
 	printValues(this->Hospitals);
 	std::cout << std::endl;
 	printValues(this->Residents);
-    //std::cout << std::endl;
-    //printValues(this->ResidentsPreference);
 }
 
-void AlgorithmProblem::printValues(std::map<int, std::vector<int>> input)
-{
-	// for (int i = 0; i < input.size(); i++)
-	// {
-	// 	for (int j = 0; j < input[i].size(); j++)
-	// 	{
-	// 		if (j != (input[i].size() - 1))
-	// 			std::cout << to_format(input[i][j]) << " ";
-	// 		else
-	// 			std::cout << to_format(input[i][j]) << std::endl;
-	// 	}
-	// }
-
+void AlgorithmProblem::printValues(std::map<int, std::vector<int>> input){
     for (auto it{input.begin()}; it != input.end(); ++it){
-        for (int j{0};  j < it->second.size();++j){
-            if (j != (it->second.size() - 1))
+        for (int j{0};  j < static_cast<int>(it->second.size());++j){
+            if (j != (static_cast<int>(it->second.size()) - 1))
 				std::cout << to_format(it->second[j]) << " ";
 			else
 				std::cout << to_format(it->second[j]) << std::endl;
